+++ src/serena/tools/memory_tools.py
@@ -49,7 +49,7 @@ class ListMemoriesTool(Tool):
         """
         List available memories. Any memory can be read using the `read_memory` tool.
         """
-        return self._to_json(self.memories_manager.list_memories())
+        return self._to_output(self.memories_manager.list_memories())
 
 
 class DeleteMemoryTool(Tool):